{memberId} 삭제
